+++ src/main/java/io/github/torand/fastersql/dialect/PostgreSqlDialect.java
@@ -65,13 +65,18 @@ public class PostgreSqlDialect implements Dialect {
     }
 
     @Override
+    public String formatToCharFunction(String operand, String format) {
+        return "to_char(" + operand + ", " + format + ")";
+    }
+
+    @Override
     public String formatSubstringFunction(String operand, int startPos, int length) {
         return "substring(" + operand + ", " + startPos + ", " + length + ")";
     }
 
     @Override
     public String formatConcatFunction(List<String> operands) {
-        throw new UnsupportedOperationException("Use the concat infix operator for PostgreSql");
+        throw new UnsupportedOperationException("PostgreSql does not support the concat() function (use the concat infix operator instead)");
     }
 
     @Override